libmythtv: Move files in HLS/ to their only users
Note that httplivestreambuffer.cpp implements a separate
parser for .m3u files from the one in m3u.cpp.
